Rename button icons to EuiButtonIcon component
